COM-2690 - Remove useless stuff in mixin
