+++ src/modules/client/mixins/subscriptionMixin.js
@@ -1,7 +1,6 @@
 import get from 'lodash/get';
 import has from 'lodash/has';
-import capitalize from 'lodash/capitalize';
-import { MONTHLY, FIXED, HOURLY, NATURE_OPTIONS, WEEKS_PER_MONTH } from '@data/constants';
+import { MONTHLY, FIXED, HOURLY, WEEKS_PER_MONTH } from '@data/constants';
 import { getLastVersion } from '@helpers/utils';
 import moment from '@helpers/moment';
 
@@ -11,44 +10,6 @@ export const subscriptionMixin = {
       subscriptions: [],
       selectedSubscription: {},
       subscriptionHistoryModal: false,
-      subscriptionsColumns: [
-        { name: 'service', label: 'Service', align: 'left', field: row => get(row, 'service.name') },
-        {
-          name: 'nature',
-          label: 'Nature',
-          align: 'left',
-          format: (value) => {
-            const nature = NATURE_OPTIONS.find(option => option.value === value);
-            return nature ? capitalize(nature.label) : '';
-          },
-          field: row => get(row, 'service.nature'),
-        },
-        {
-          name: 'unitTTCRate',
-          label: 'Prix unitaire TTC',
-          align: 'center',
-          field: row => row.unitTTCRate && `${this.formatNumber(row.unitTTCRate)}€`,
-        },
-        {
-          name: 'weeklyHours',
-          label: 'Volume horaire hebdomadaire estimatif',
-          align: 'center',
-          field: row => (row.weeklyHours ? `${row.weeklyHours}h` : ''),
-        },
-        {
-          name: 'weeklyCount',
-          label: 'Nombre d\'interventions hebdomadaire estimatif',
-          align: 'center',
-          field: row => (row.weeklyCount || ''),
-        },
-        {
-          name: 'weeklyRate',
-          label: 'Coût hebdomadaire TTC *',
-          align: 'center',
-          field: row => `${this.formatNumber(this.computeWeeklyRate(row, this.getMatchingFunding(row)).total)}€`,
-        },
-        { name: 'actions', label: '', align: 'left', field: '_id' },
-      ],
       subscriptionHistoryColumns: [
         {
           name: 'createdAt',
@@ -93,9 +54,6 @@ export const subscriptionMixin = {
     };
   },
   methods: {
-    formatNumber (number) {
-      return parseFloat(Math.round(number * 100) / 100).toFixed(2);
-    },
     computeWeeklyRate (subscription, funding) {
       let weeklyRate = subscription.weeklyHours
         ? subscription.unitTTCRate * subscription.weeklyHours
